src/requirements.txt
python-telegram-bot==20.7
python-dateutil==2.8.2
APScheduler==3.10.4 